Update s3 sync according to the new events naming
