+++ tests/functional/configure/test_sso.py
@@ -1,0 +1,22 @@
+# Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You
+# may not use this file except in compliance with the License. A copy of
+# the License is located at
+#
+#     http://aws.amazon.com/apache2.0/
+#
+# or in the "license" file accompanying this file. This file is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
+# ANY KIND, either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+from awscli.testutils import BaseAWSCommandParamsTest
+
+
+class TestConfigureSSO(BaseAWSCommandParamsTest):
+    prefix = 'configure sso'
+
+    def test_sso_logout_do_not_exist(self):
+        cmdline = self.prefix + ' logout'
+        _, err, _ = self.run_cmd(cmdline, 252)
+        self.assertIn('Unknown options', err)
