+++ tests/unit/customizations/s3/test_s3.py
@@ -28,7 +28,7 @@ class AWSInitializeTest(unittest.TestCase):
         awscli_initialize(self.cli)
         reference = []
         reference.append("building-command-table.main")
-        reference.append("building-command-table.sync")
+        reference.append("building-command-table.s3_sync")
         for arg in self.cli.register.call_args_list:
             self.assertIn(arg[0][0], reference)
